move speedscope references to its own file with license header
