+++ static/app/utils/profiling/flamegraph.ts
@@ -35,10 +35,6 @@ function makeTreeSort(sortFn: (a: CallTreeNode, b: CallTreeNode) => number) {
 const alphabeticTreeSort = makeTreeSort(sortAlphabetically);
 const leftHeavyTreeSort = makeTreeSort(sortByTotalWeight);
 
-// Intermediary flamegraph data structure for rendering a profile. Constructs a list of frames from a profile
-// and appends them to a virtual root. Taken mostly from speedscope with a few modifications. This should get
-// removed as we port to our own format for profiles. The general idea is to iterate over profiles while
-// keeping an intermediary stack so as to resemble the execution of the program.
 export class Flamegraph {
   profile: Profile;
   frames: ReadonlyArray<FlamegraphFrame> = [];
@@ -167,9 +163,9 @@ export class Flamegraph {
       0
     );
 
-    this.root.node.addToTotalWeight(weight);
+    this.root.node.totalWeight += weight;
     this.root.end = this.root.start + weight;
-    this.root.frame.addToTotalWeight(weight);
+    this.root.frame.totalWeight += weight;
   }
 
   buildCallOrderChart(profile: Profile): FlamegraphFrame[] {
