+++ static/app/utils/profiling/guards/profile.tsx
@@ -30,41 +30,3 @@ export function isSentrySampledProfile(
     'frames' in profile.profile
   );
 }
-
-export function isChromeTraceObjectFormat(
-  profile: any
-): profile is ChromeTrace.ObjectFormat {
-  return typeof profile === 'object' && 'traceEvents' in profile;
-}
-
-// We check for the presence of at least one ProfileChunk event in the trace
-export function isChromeTraceArrayFormat(
-  profile: any
-): profile is ChromeTrace.ProfileType {
-  return (
-    Array.isArray(profile) && profile.some(p => p.ph === 'P' && p.name === 'ProfileChunk')
-  );
-}
-
-// Typescript uses only a subset of the event types (only B and E cat),
-// so we need to inspect the contents of the trace to determine the type of the profile.
-// The TS trace can still contain other event types like metadata events, meaning we cannot
-// use array.every() and need to check all the events to make sure no P events are present
-export function isTypescriptChromeTraceArrayFormat(
-  profile: any
-): profile is ChromeTrace.ArrayFormat {
-  return (
-    Array.isArray(profile) &&
-    !profile.some(p => p.ph === 'P' && p.name === 'ProfileChunk')
-  );
-}
-
-export function isChromeTraceFormat(
-  profile: Readonly<any>
-): profile is ChromeTrace.ArrayFormat {
-  return (
-    isTypescriptChromeTraceArrayFormat(profile) ||
-    isChromeTraceObjectFormat(profile) ||
-    isChromeTraceArrayFormat(profile)
-  );
-}
