+++ static/app/utils/profiling/profile/importProfile.tsx
@@ -4,17 +4,13 @@ import {Transaction} from '@sentry/types';
 import {Image} from 'sentry/types/debugImage';
 
 import {
-  isChromeTraceFormat,
-  isChromeTraceObjectFormat,
   isEventedProfile,
   isJSProfile,
   isSampledProfile,
   isSchema,
   isSentrySampledProfile,
-  isTypescriptChromeTraceArrayFormat,
 } from '../guards/profile';
 
-import {parseTypescriptChromeTraceArrayFormat} from './chromeTraceProfile';
 import {EventedProfile} from './eventedProfile';
 import {JSSelfProfile} from './jsSelfProfile';
 import {Profile} from './profile';
@@ -60,14 +56,6 @@ export function importProfile(
         transaction.setTag('profile.type', 'js-self-profile');
       }
       return importJSSelfProfile(input, traceID, {transaction, type});
-    }
-
-    if (isChromeTraceFormat(input)) {
-      // In some cases, the SDK may return transaction as undefined and we dont want to throw there.
-      if (transaction) {
-        transaction.setTag('profile.type', 'chrometrace');
-      }
-      return importChromeTrace(input, traceID, {transaction, type});
     }
 
     if (isSentrySampledProfile(input)) {
@@ -119,22 +107,6 @@ function importJSSelfProfile(
       durationNS: profile.duration,
     },
   };
-}
-
-function importChromeTrace(
-  input: ChromeTrace.ProfileType,
-  traceID: string,
-  options: ImportOptions
-): ProfileGroup {
-  if (isChromeTraceObjectFormat(input)) {
-    throw new Error('Chrometrace object format is not yet supported');
-  }
-
-  if (isTypescriptChromeTraceArrayFormat(input)) {
-    return parseTypescriptChromeTraceArrayFormat(input, traceID, options);
-  }
-
-  throw new Error('Failed to parse trace input format');
 }
 
 function importSentrySampledProfile(
@@ -248,11 +220,7 @@ export function importSchema(
 }
 
 function importSingleProfile(
-  profile:
-    | Profiling.EventedProfile
-    | Profiling.SampledProfile
-    | JSSelfProfiling.Trace
-    | ChromeTrace.ProfileType,
+  profile: Profiling.EventedProfile | Profiling.SampledProfile | JSSelfProfiling.Trace,
   frameIndex: ReturnType<typeof createFrameIndex>,
   {transaction, type, profileIds}: ImportOptions
 ): Profile {
