+++ static/app/utils/profiling/renderers/gridRenderer.spec.tsx
@@ -1,10 +1,11 @@
 import {LightFlamegraphTheme} from 'sentry/utils/profiling/flamegraph/flamegraphTheme';
 import {Rect, transformMatrixBetweenRect} from 'sentry/utils/profiling/gl/utils';
 import {
-  computeInterval,
   getIntervalTimeAtX,
   GridRenderer,
 } from 'sentry/utils/profiling/renderers/gridRenderer';
+
+import {computeInterval} from '../speedscope';
 
 describe('getIntervalTimeAtX', () => {
   beforeEach(() => {
@@ -81,7 +82,7 @@ describe('computeInterval', () => {
 
     const logicalToConfig = transformMatrixBetweenRect(logicalSpace, configView);
 
-    expect(computeInterval(configView, logicalToConfig)).toEqual([
+    expect(computeInterval(configView, logicalToConfig, getIntervalTimeAtX)).toEqual([
       0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100,
     ]);
   });
@@ -96,7 +97,7 @@ describe('computeInterval', () => {
 
     const logicalToConfig = transformMatrixBetweenRect(logicalSpace, configView);
 
-    expect(computeInterval(configView, logicalToConfig)).toEqual([
+    expect(computeInterval(configView, logicalToConfig, getIntervalTimeAtX)).toEqual([
       50, 60, 70, 80, 90, 100,
     ]);
   });
@@ -111,7 +112,7 @@ describe('computeInterval', () => {
 
     const logicalToConfig = transformMatrixBetweenRect(logicalSpace, configView);
 
-    expect(computeInterval(configView, logicalToConfig)).toEqual([
+    expect(computeInterval(configView, logicalToConfig, getIntervalTimeAtX)).toEqual([
       60, 65, 70, 75, 80, 85, 90, 95, 100,
     ]);
   });
@@ -126,7 +127,7 @@ describe('computeInterval', () => {
 
     const logicalToConfig = transformMatrixBetweenRect(logicalSpace, configView);
 
-    expect(computeInterval(configView, logicalToConfig)).toEqual([
+    expect(computeInterval(configView, logicalToConfig, getIntervalTimeAtX)).toEqual([
       80, 82, 84, 86, 88, 90, 92, 94, 96, 98, 100,
     ]);
   });
@@ -141,7 +142,7 @@ describe('computeInterval', () => {
 
     const logicalToConfig = transformMatrixBetweenRect(logicalSpace, configView);
 
-    expect(computeInterval(configView, logicalToConfig)).toEqual([
+    expect(computeInterval(configView, logicalToConfig, getIntervalTimeAtX)).toEqual([
       0, 0.5, 1, 1.5, 2, 2.5, 3,
     ]);
   });
